Export accesscontrol types
We want to be able to use interfaces such as
[Permission](https://onury.io/accesscontrol/?api=ac#AccessControl~Permission)
in our typedefs, so if we use `InjectRolesBuilder() ac: RolesBuilder` we
can say `let perm: Permission = this.ac.can(roles).readAny('photos')`.

Alternatives:
Just use type inference: without the `: Permission`,  `perm` is still
inferred as `Permission` type.  But what if we want to be explicit for
readability?

Separately add `accesscontrol` to our `package.json` and import
`Permission` separately.  This is fine (and what we'll do if this PR
doesn't get accepted), but why not just give us these?

Is it because you rename `AccessControl`→`RoleBuilder`, and you don't
want to add confusion by having both types available?